Fix for RT-21830: Improve Color startup by shortening the rgb calls for the predefined colors.
The fix was to add a private constructor which takes 3 params instead of 4, all floats. (The 3 vs. 4 params makes it easier to make sure that we're only calling this constructor internally for the static colors). The static colors were converted over to use this constructor, and using pre-computed r, g, b float values (vs. the old approach of / 255.0 for each color at runtime).

As this was a "Medium" priorty issue, probably actually a "Minor", I don't expect major gains from this. Just a little piece of salami.