Fix for RT-16621: Selected text is blue even focus is on another control.
The fix is simply to style the TextArea such that it appears gray instead of blue when not focused. This required a fix to the css file, as well as to the TextInputControlSkin class. There was a bug there where it wouldn't notify the TextAreaSkin (via protected methods) when the highlight fill, or highlight text fill changed, and as a result, the CSS wasn't being applied to the TextAreaSkin.