[RT-23229] - DFB: Command prompt appears in images in Controls test suite
Problem:
Command prompt cursor is been set to the left top corner of the screen and blinks while fx application run in DFB

Analysis:
For some reason there is an overlay between the command prompt console of the device and the framebuffer useb by FB.
I assume this have something to do with the platform windowing, but I couldn't find a way to interact with it.

Solution:
This is not perfect, but manually turning the blinking off, by writing '1' or '0' to "/sys/class/graphics/fbcon/cursor_blink", solves the problem.
A side effect of this that rendering improves and become less jerky.